pyproject.toml
[build-system]
requires=["setuptools>=42"] 
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-flow-builder"
version = "0.0.1"
description = "DiffuseFlowBuilder is a versatile YAML-based configuration repository designed to streamline the creation and execution of stable diffusion pipelines. This powerful tool empowers users to effortlessly define and manage pipelines for essential diffusion operations, including text-to-image conversion, in-painting, and super-resolution."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3"
]
dependencies = [
    "torch",
    "torchvision",
    "transformers[torch]",
    "diffusers[torch]",
    "compel",
    "Pillow",
    "tqdm",
    "pytest"
]

[project.urls]
"Homepage" = "https://github.com/LeoBaro/DiffuseFlowBuilder"
"Bug Tracker" = "https://github.com/LeoBaro/DiffuseFlowBuilder/issues"

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "tests*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "tests*"]

[project.scripts]
